perf: make this._json calculation lazy
I've been doing some profiling of js-ipfs and where there are large amounts
of DAG operations being performed (importing all of npm to ipfs, for
example) we spend about 1.8-2.5% of our time in the creation of this._json
objects, even though they are not always used, so this PR makes them be
created when first requested.

It also passes the `this._json` object through `Object.freeze` becase otherwise
we expose a mutable object to the world - it seems odd that we've taken pains
to ensure that the properties of a DAGNode/DAGLink are immutable, then expose
objects whose properties are not immutable.

Finally it makes more use of the `cids` module to parse CIDs from multihashes
as the `cids` module respects CIDv1 versions/codecs if present whereas the
`multihashes` module does not and is only really suitable for use with CIDv0.