perf: remove named links from object
This change was motivated by performance issues, but it also
really is cleaner to keep the IPLD objects lean and not cluttered
with IPFS related things.

BREAKING CHANGE: named links are no longer part of an object

Access to named links is only possible with calling `resolve()`.
Hence they are also not part of `tree()` anymore.

Named links are a feature of IPFS and only supported for
backwards compatibility, they are not really part of IPLD.