fix: strip control characters from link names
In the case where we receive garbage or malicious strings that contain
characters with a utf code less than 31, strip those character codes
from the link name.

We keep the garbage string around though, so we can de-serialize and
reserialize a node without changing it's CID, though we do not expose
that string to the caller through documented properties.